Chapter 4.1 Descriptor layout and buffer
Completed the first section of chapter 4 (Uniform buffers). Learned
about descriptors with uniform buffers object (UBO) as an example. Updated
the vertex shader to handle the UBO and created a descriptor layout to
specify what resources are going to be accessed by the pipeline. Lastly
I setup the uniform buffers, and created a function used to update a uniform
buffer every frame.